refactor: move cleaner inline tests to separate file, remove dead IMAdapter methods
- Move inline #[cfg(test)] tests from cleaner.rs to cleaner_tests.rs
- Add #[cfg(test)] to normalized_tests and plugin_tests module declarations
- Remove close_inbound()/close_outbound() dead methods from IMAdapter trait

Source: user
Type: refactor
