+++ src/im_adapter/mod.rs
@@ -6,9 +6,11 @@
 pub mod code_block;
 pub mod error;
 pub mod normalized;
+#[cfg(test)]
 pub mod normalized_tests;
 pub mod platforms;
 pub mod plugin;
+#[cfg(test)]
 pub mod plugin_tests;
 pub mod streaming;
 
@@ -56,21 +58,5 @@ pub trait IMAdapter: Send + Sync {
         _root_id: Option<&str>,
     ) -> Result<(), AdapterError> {
         Err(AdapterError::UnsupportedOperation)
-    }
-
-    /// Close inbound connections (e.g. unsubscribe webhook, disconnect WebSocket).
-    ///
-    /// Called during daemon Phase 1 (inbound shutdown) to stop accepting new
-    /// messages from the platform. Default implementation is a no-op.
-    async fn close_inbound(&self) -> Result<(), AdapterError> {
-        Ok(())
-    }
-
-    /// Close outbound connections (e.g. drain send queue, disconnect API client).
-    ///
-    /// Called during daemon Phase 5 (outbound shutdown) to stop sending
-    /// messages to the platform. Default implementation is a no-op.
-    async fn close_outbound(&self) -> Result<(), AdapterError> {
-        Ok(())
     }
 }
